+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
+## [6.3.0] 2025-09-06
+
 - New command [hardis:org:diagnose:unsecured-connected-apps](https://sfdx-hardis.cloudity.com/hardis/org/diagnose/unsecured-connected-apps/) to detect Unsecured Connected Apps
 - Add documentation about Packages installation
 - Update Azure Pipelines integration documentation